Тесты разделены на два класса
